refac(seller_lead): extracting functions related with duplicity checking to its own module
